refactor(web,mobile): consomme @kinhale/sync/client, supprime duplication
Troisième et dernière étape de KIN-039. Les apps web et mobile consomment
désormais le hook et la dérivation groupKey depuis @kinhale/sync/client,
point d'entrée unique mutualisé. Les cinq fichiers dupliqués sont supprimés
de chaque app (group-key.ts, useRelaySync.ts, RelaySyncBootstrap.tsx +
deux tests), soit ~950 lignes de code redondant retirées.

Chaque app conserve un wrapper fin dans src/lib/sync/index.ts qui injecte
ses dépendances plateforme concrètes :
- Hooks Zustand (useAuthStore, useDocStore) — diffèrent par leur storage
  backend (IndexedDB persist web vs AsyncStorage mobile)
- Factory WebSocket plateforme (createRelayClient web DOM vs RN natif)
- La dérivation groupKey est partagée (même implémentation Argon2id)

RelaySyncBootstrap (composant 3 lignes) est inline dans le même index.ts
des deux apps plutôt que mutualisé : le pragma 'use client' requis par
Next.js diffère structurellement de la version RN, et centraliser un
composant aussi trivial serait du sur-nivelage sans gain concret.

Config test :
- apps/web/jest.config.ts : moduleNameMapper pour @kinhale/sync/client
  vers packages/sync/src/client/index.ts
- apps/mobile/jest.config.js : idem mapper + nouveau stub sync-client.ts
  dans __mocks__ (useRelaySync retourne {connected: false}, l'app mobile
  ne testait pas la logique du hook de toute façon)

Aucun changement de comportement runtime : mêmes blobs émis, mêmes
messages WS, même gestion du cursor, mêmes erreurs silenciées. Zero-
knowledge préservé (groupKey toujours en mémoire uniquement).

Prettier a reformaté packages/sync/src/client/__tests__/useRelaySync.test.tsx
(inclus dans ce commit puisque détecté lors du pnpm format:check après la
migration).

Refs: KIN-039
Closes: #190
