+++ apps/mobile/src/providers/index.tsx
@@ -3,7 +3,7 @@ import { TamaguiProvider } from 'tamagui';
 import { I18nextProvider } from 'react-i18next';
 import config from '../lib/tamagui.config';
 import i18n from '../lib/i18n';
-import { RelaySyncBootstrap } from '../lib/sync/RelaySyncBootstrap';
+import { RelaySyncBootstrap } from '../lib/sync';
 
 export function Providers({ children }: { children: ReactNode }): JSX.Element {
   return (
