+++ apps/web/jest.config.ts
@@ -12,6 +12,7 @@ const customConfig: Config = {
     '^@kinhale/i18n$': '<rootDir>/../../packages/i18n/src/index.ts',
     '^@kinhale/crypto$': '<rootDir>/../../packages/crypto/src/index.ts',
     '^@kinhale/sync$': '<rootDir>/../../packages/sync/src/index.ts',
+    '^@kinhale/sync/client$': '<rootDir>/../../packages/sync/src/client/index.ts',
   },
   transformIgnorePatterns: [
     '/node_modules/(?!(react-native-web|@tamagui|tamagui|@react-native|idb|fake-indexeddb)/)/',
